pyproject.toml
[tool.poetry]
name = "gwinstek-api-client"
version = "0.1.0"
description = "Api client for power supply"
authors = ["Roman Popov <i@romanpopoff.ru>"]
readme = "README.md"
packages = [{include = "gwinstek_api_client"}]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.3.1"
uvicorn = "^0.22.0"
fastapi = "^0.95.2"
pyvisa-py = "^0.7.0"
flake8 = "^6.0.0"
pytest-asyncio = "^0.21.0"
httpx = "^0.24.1"
unicorn = "^2.0.1.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
